Fixes#9576: Data insights page empty and reindexing fails
